fix: 노드 추상화 없이 각자 구현 -> 스타일 파일 분리, preview 노드 분리
