--- src/common/components/OkrTree/Nodes/PreviewOkrNodes.tsx
@@ -1,1 +1,0 @@
-export const PreviewTaskNodes = () => {};
